User Update API fixed (Show whole User data in Response) (dev -> Dev)

1. User Update API fixed (Show whole User data in Response)
2. UserProfileImages img_profile field changed for testing2

+++ app/members/models.py
@@ -155,6 +155,9 @@ def remove_file_from_storage(sender, instance, using, **kwargs):
     #     img_url = instance.img_profile.url
     #     print(img_url)
 
-    instance.img_profile.delete(save=False)
-    instance.img_profile_28.delete(save=False)
-    instance.img_profile_225.delete(save=False)
+    if instance.img_profile:
+        instance.img_profile.delete(save=False)
+    if instance.img_profile_28:
+        instance.img_profile_28.delete(save=False)
+    if instance.img_profile_225:
+        instance.img_profile_225.delete(save=False)
